# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: DClaw Video API
  version: 1.0.0
paths:
...
  /api/v1/projects:
    post:
      summary: Create project from script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                script_text: { type: string }
                template: { type: string, enum: [youtube_explainer, product_demo, social_reel] }
                voice_profile_id: { type: string }
                character_id: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    get:
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /api/v1/projects/{id}:
    get:
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
components:
  schemas:
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        script_text: { type: string }
        status: { type: string, enum: [draft, storyboard, rendering, done] }
        video_url: { type: string, nullable: true }
        duration: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
...